+++ packages/signals_core/lib/src/core/computed.dart
@@ -201,6 +201,9 @@ class _Computed<T> implements Computed<T>, _Listenable {
         }
       }
     }
+    if (autoDispose && _allTargets.isEmpty) {
+      dispose();
+    }
   }
 
   @override
@@ -238,8 +241,13 @@ class _Computed<T> implements Computed<T>, _Listenable {
   @override
   T get value {
     if (disposed) {
-      throw SignalsReadAfterDisposeError(this);
+      if (kDebugMode) {
+        print(
+            'computed warning: [$globalId|$debugLabel] has been read after disposed');
+      }
+      return this._value;
     }
+
     if ((_flags & RUNNING) != 0) {
       _cycleDetected();
     }
@@ -292,19 +300,22 @@ class _Computed<T> implements Computed<T>, _Listenable {
 
   @override
   void dispose() {
+    if (disposed) return;
     for (final cleanup in _disposeCallbacks) {
       cleanup();
     }
     _disposeCallbacks.clear();
     _flags |= DISPOSED;
-    if (_node != null) _unsubscribe(_node!);
-    _value = _initialValue;
-    _previousValue = _initialValue;
     disposed = true;
   }
 
   @override
   T get initialValue => _initialValue;
+
+  void reset([T? value]) {
+    _value = value ?? _initialValue;
+    _previousValue = value ?? _initialValue;
+  }
 }
 
 typedef ComputedCallback<T> = T Function();